DataBase/requirements.txt
python-dotenv==1.0.0
firebase-admin==6.3.0
requests==2.31.0
pandas==2.1.4
geopy==2.4.1